Fix Codecov integration
Codecov is not working if CI does a Release build. This commit updates
the CI config so that it will only do a Debug build. It also updates
the versions of cocoapods and Xcode to use. Additionally, we now use
the Codecov -J option so that coverage is only reported for the
framework sources, not for the test sources. Lastly, we now inhibit
geos warnings to shorten build logs and reduce the warning noise when
developing GEOSwift.